Move pathfinding to an appropriate location, fix compilation error
