+++ src/main/java/net/tridentsdk/server/entity/living/ai/pathfind/Pathfinder.java
@@ -15,18 +15,20 @@
  * limitations under the License.
  */
 
-package net.tridentsdk.server.entity.ai.pathfinder;
+package net.tridentsdk.server.entity.living.ai.pathfind;
 
-import net.tridentsdk.Coordinates;
+import net.tridentsdk.Position;
+import net.tridentsdk.base.Block;
 import net.tridentsdk.base.Substance;
-import net.tridentsdk.base.Tile;
 import net.tridentsdk.entity.Entity;
 import net.tridentsdk.util.Vector;
 
+import javax.annotation.concurrent.NotThreadSafe;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 
+@NotThreadSafe
 public class Pathfinder {
     private final Entity entity;
     private final Node start;
@@ -37,9 +39,9 @@ public class Pathfinder {
     private final HashSet<Node> closedList = new HashSet<>();
     private Path result;
 
-    public Pathfinder(Entity entity, Coordinates target, double range) {
+    public Pathfinder(Entity entity, Position target, double range) {
         this.entity = entity;
-        this.start = new Node(null, entity.getLocation());
+        this.start = new Node(null, entity.location());
         this.end = new Node(null, target);
         this.range = range;
     }
@@ -87,7 +89,7 @@ public class Pathfinder {
             }
         }
 
-        if(end.getParent() != null) {
+        if(end.parent() != null) {
             this.result = new Path(end);
         }
     }
@@ -128,13 +130,13 @@ public class Pathfinder {
     }
 
     private boolean canWalkOn(Node node) {
-        Tile tile = entity.getWorld().tileAt(Coordinates.create(entity.getWorld(), node.getX(), node.getY(), node.getZ()));
-        return tile.substance().isSolid();
+        Block block = entity.world().blockAt(Position.create(entity.world(), node.x(), node.y(), node.z()));
+        return block.substance().isSolid();
     }
 
     private boolean canWalkThrough(Node node) {
-        Tile tile = entity.getWorld().tileAt(Coordinates.create(entity.getWorld(), node.getX(), node.getY(), node.getZ()));
-        return canWalkThrough(tile.substance()) && canWalkThrough(tile.relativeTile(new Vector(0, 1, 0)).substance());
+        Block block = entity.world().blockAt(Position.create(entity.world(), node.x(), node.y(), node.z()));
+        return canWalkThrough(block.substance()) && canWalkThrough(block.relativeBlock(new Vector(0, 1, 0)).substance());
     }
 
     private boolean canWalkThrough(Substance type) {
